[rust png] Fixing forward implicit int cast.
https://review.skia.org/1058817 accidentally introduced an implicit
integer cast.  This CL fixes this, which should unblock the Chromium
roll.

Bug: chromium:444045690
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1059856
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Auto-Submit: Łukasz Anforowicz <lukasza@google.com>

+++ include/private/SkEncodedInfo.h
@@ -243,7 +243,7 @@ private:
         , fHdrMetadata(hdrMetadata)
     {}
 
-    static void VerifyColor(Color color, Alpha alpha, uint8_t bitsPerComponent) {
+    static void VerifyColor(Color color, Alpha alpha, int bitsPerComponent) {
         // Avoid `-Wunused-parameter` warnings on non-debug builds.
         std::ignore = alpha;
         std::ignore = bitsPerComponent;
